+++ src/Event/ListBeforeRenderItemEvent.php
@@ -9,34 +9,32 @@
 namespace HeimrichHannot\ListBundle\Event;
 
 use HeimrichHannot\ListBundle\Item\ItemInterface;
+use HeimrichHannot\ListBundle\ListConfiguration\ListConfiguration;
 use Symfony\Contracts\EventDispatcher\Event;
 
 class ListBeforeRenderItemEvent extends Event
 {
-    const NAME = 'huh.list.event.item_before_render';
-    /**
-     * @var string
-     */
-    protected $templateName;
-    /**
-     * @var array
-     */
-    protected $templateData;
-    /**
-     * @var ItemInterface
-     */
-    protected $item;
+    public const NAME = 'huh.list.event.item_before_render';
 
-    /**
-     * ListBeforeRenderItemEvent constructor.
-     *
-     * @param $templateName
-     */
-    public function __construct(string $templateName, array $templateData, ItemInterface $item)
+    protected string $templateName;
+
+    protected array $templateData;
+
+    protected ItemInterface   $item;
+
+    private ListConfiguration $listConfiguration;
+
+    public function __construct(
+        string $templateName,
+        array $templateData,
+        ItemInterface $item,
+        ListConfiguration $listConfiguration
+    )
     {
         $this->templateName = $templateName;
         $this->templateData = $templateData;
         $this->item = $item;
+        $this->listConfiguration = $listConfiguration;
     }
 
     public function getTemplateName(): string
@@ -67,5 +65,10 @@ class ListBeforeRenderItemEvent extends Event
     public function setItem(ItemInterface $item): void
     {
         $this->item = $item;
+    }
+
+    public function getListConfiguration(): ListConfiguration
+    {
+        return $this->listConfiguration;
     }
 }